effects/overview: Allow navigation between virtual desktops
Previously when you clicked on another virtual desktop on the desktop bar
it would switch to that virtual desktop but end the effect.
Now when you click on another virtual desktop you stay on the overview
effect, if you click on the same virtual desktop you are already on
then the effects ends.

This commit provides a nice way to navigate between your virtual
desktop without leaving the overview effect. The previous behaviour
was quite intrusive and not intuitive.

BUG: 448668