effects/overview: Place panels behind blurred overlay
Since panels are visible currently, you can be tricked into thinking
that they are interactive, but they are not due to various technical
details.

With the current design of plasma, the compositor and the desktop shell
live in two different processes. In theory, it will result in a more
robust desktop; in practice, implementing features such as interactive
panels in the overview effect become difficult.

On Wayland, panel interactivity issues are solvable, but we would still
have other issues, e.g. panel popups not being animated as during normal
operation mode, etc.

For the time being, this change puts panels behind blurred overlay to
avoid confusion.

BUG: 444274