autotests: Fix QuickTilingTest
AbstractClient::geometryRestore() is no longer initialized when the
window is mapped so XdgToplevelClient can send a configure event with
0x0 size if user tries to restore initially maximized window.

However, it broke AbstractClient::setQuickTileMode().

This change intends to fix it (and relevant test) by making maximize
quick tile mode compute geometry restore.

If user has dragged the window to a screen edge in order to maximize it,
geometryRestore() will be set to the geometry that the window had when
user started moving it; otherwise use the current geometry.